moved all configuration routines to config.js, and established getter/setter mechanism to prevent circular dependency
